pyproject.toml
[tool.poetry]
name = "site"
version = "0.1.0"
description = ""
authors = ["DouglasKrouth <douglaskrouth@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
spotipy = "^2.23.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
